requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-postgres==5.5.0
apache-airflow-providers-http==4.5.0
requests>=2.28.0
python-dotenv>=0.21.0
psycopg2-binary>=2.9.0

